+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Not released
 
+- Default props of core components to meet the common use case [#552](https://github.com/CartoDB/carto-react/pull/552/)
 - Chip component adapted to the new design system [#551](https://github.com/CartoDB/carto-react/pull/551/)
 - Design changes in UploadField component [#548](https://github.com/CartoDB/carto-react/pull/548/)
 - Divider component adapted to the new design system [#546](https://github.com/CartoDB/carto-react/pull/546/)